chore(watch): address Copilot review — plurals, Int64 cast, comment
- Localizable.xcstrings: watch.today.countMixed now uses a plural
  substitution on the "others" arg so DE/ES pick correct singular
  vs plural word forms (weiteres/weitere, otra/otras) instead of
  always-plural. English reads naturally too (1 other / 3 others).
- MainView.swift: cast Int → Int64 when passing to String(format:)
  with %lld to match the format specifier exactly.
- index.swift: reconcile contradictory comments — both now state
  the write takes effect on the next cold launch.

+++ apps/mobile/targets/watch/Views/MainView.swift
@@ -111,7 +111,7 @@ struct MainView: View {
                     "watch.today.count",
                     comment: "Today summary when only beers were logged. %lld is drink count."
                 ),
-                viewModel.drinkCount
+                Int64(viewModel.drinkCount)
             )
         }
         return String(
@@ -119,8 +119,8 @@ struct MainView: View {
                 "watch.today.countMixed",
                 comment: "Today summary with beers and other drinks. %1$lld is beers, %2$lld is others."
             ),
-            viewModel.beerCount,
-            others
+            Int64(viewModel.beerCount),
+            Int64(others)
         )
     }
